+++ packages/js/src/make_script_data/output_script_data_schema.ts
@@ -5,7 +5,10 @@ type CommonScriptData = {
    * `string` is the actual key, number(nullable) is the index of the key in the compiled `list`
    */
   krama_text_arr: [krama_key: string, list_arr_ref: number | null][];
-  /** To be used for binary search on the `krama_key_map` */
+  /** To be used for binary search on the `krama_key_map`
+   * Although now we are not using this, but lets not remove it yet.
+   * As it can be useful in getting the order of krama texts, like a for a Typing Helper
+   */
   krama_text_arr_index: number[];
   /** `string` used for search. Usable for Duplication Remova, Fallback Behaviour
    * and for `other` scripts type. Like Normal