+++ packages/rust/src/macros.rs
@@ -1,7 +1,0 @@
-macro_rules! is_script_tamil_ext {
-  ($var:expr) => {
-    $var == "Tamil-Extended"
-  };
-}
-
-pub(crate) use is_script_tamil_ext;